Commit before upgradde HAL

+++ CM7/Peripheral/Inc/cis.h
@@ -41,7 +41,7 @@ typedef enum
 #define CIS_BLUE_LANE_OFFSET  		((CIS_LANE_SIZE * 2)) + (CIS_START_OFFSET)
 
 void cis_init(void);
-void cis_getRAWImage(float32_t* cisDataCpy_f32, uint16_t overSampling);
+void cis_getRAWImage(float32_t* cisDataCpy_f32, uint8_t overSampling);
 void cis_convertRAWImageToFloatArray(float32_t* cisDataCpy_f32, struct RAWImage* RAWImage);
 void cis_imageProcess_2(int32_t *cis_buff);
 void cis_imageProcess(float32_t* cisDataCpy_f32, struct packet_Image *imageBuffers);
